feat: Continue orderTicket flux implementation

+++ src/main/java/com/muvent/api/domain/coupon/Coupon.java
@@ -18,7 +18,7 @@ import java.time.LocalDate;
 public class Coupon extends BaseEntity {
 
     private String code;
-    private Integer discount;
+    private Long discount;
     private LocalDate valid;
 
     @ManyToOne
